LayoutTests/css3/color-filters/color-filter-exposed-if-disabled-expected.txt

Tests that -apple-color-filter is not exposed when the feature is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS '-apple-color-filter' in document.documentElement.style is false
PASS document.documentElement.style['-apple-color-filter'] is undefined.
PASS document.documentElement.style.getPropertyValue('-apple-color-filter') is ""
PASS document.documentElement.style.getPropertyPriority('-apple-color-filter') is ""
PASS document.documentElement.style.getPropertyCSSValue('-apple-color-filter') is null
PASS document.documentElement.style.removeProperty('-apple-color-filter') is ""
PASS document.documentElement.style.getPropertyValue('-apple-color-filter') is ""
PASS document.documentElement.style.setProperty('-apple-color-filter', 'contrast(1)', '') is undefined.
PASS document.documentElement.style.getPropertyValue('-apple-color-filter') is ""
PASS '-apple-color-filter' in getComputedStyle(document.documentElement) is false
PASS getComputedStyle(document.documentElement)['-apple-color-filter'] is undefined.
PASS getComputedStyle(document.documentElement).getPropertyValue('-apple-color-filter') is ""
PASS getComputedStyle(document.documentElement).getPropertyPriority('-apple-color-filter') is ""
PASS getComputedStyle(document.documentElement).getPropertyCSSValue('-apple-color-filter') is null
PASS getComputedStyle(document.documentElement).removeProperty('-apple-color-filter') threw exception NoModificationAllowedError: The object can not be modified..
PASS document.documentElement.style.getPropertyValue('-apple-color-filter') is ""
PASS getComputedStyle(document.documentElement).setProperty('-apple-color-filter', 'contrast(1)', '') threw exception NoModificationAllowedError: The object can not be modified..
PASS document.documentElement.style.getPropertyValue('-apple-color-filter') is ""
PASS 'AppleColorFilter' in document.documentElement.style is false
PASS document.documentElement.style['AppleColorFilter'] is undefined.
PASS 'AppleColorFilter' in getComputedStyle(document.documentElement) is false
PASS getComputedStyle(document.documentElement)['AppleColorFilter'] is undefined.
PASS CSS.supports('-apple-color-filter: contrast(1)') is false
PASS CSS.supports('-apple-color-filter: inherit') is false
PASS CSS.supports('-apple-color-filter', 'contrast(1)') is false
PASS CSS.supports('-apple-color-filter', 'inherit') is false
PASS getComputedStyle(document.getElementById('test1')).color is 'rgb(0, 128, 128)'
PASS getComputedStyle(document.getElementById('test2')).color is 'rgb(0, 128, 128)'
PASS successfullyParsed is true

TEST COMPLETE
